Cleanup literal type
